feat: Add current version to UI Co-authored-by: Ray <rayokamoto@users.noreply.github.com> - Add current version to the bottom of the popup - Refactor checkForNewVersion to be reusable
