+++ src/popup.html
@@ -51,6 +51,9 @@
         <div id="download-box">
             <button id="get-ical">Download iCal file</button>
         </div>
+        <div id="footer">
+            <span>Version: </span><span id="ver-text"></span>
+        </div>
     </div>
 
     <script src="./main.js"></script>
